+++ instrumentation/github.com/gorilla/mux/otelmux/mux.go
@@ -78,13 +78,6 @@ type recordingResponseWriter struct {
 	status  int
 }
 
-func (h *recordingResponseWriter) Hijack() (net.Conn, *bufio.ReadWriter, error) {
-	if hijacker, ok := h.writer.(http.Hijacker); ok {
-		return hijacker.Hijack()
-	}
-	return nil, nil, fmt.Errorf("underlying ResponseWriter does not support hijacking")
-}
-
 var rrwPool = &sync.Pool{
 	New: func() interface{} {
 		return &recordingResponseWriter{}
@@ -115,6 +108,13 @@ func getRRW(writer http.ResponseWriter) *recordingResponseWriter {
 		},
 	})
 	return rrw
+}
+
+func (h *recordingResponseWriter) Hijack() (net.Conn, *bufio.ReadWriter, error) {
+	if hijacker, ok := h.writer.(http.Hijacker); ok {
+		return hijacker.Hijack()
+	}
+	return nil, nil, fmt.Errorf("underlying ResponseWriter does not support hijacking")
 }
 
 func putRRW(rrw *recordingResponseWriter) {